Rename the 'key' column to DBInfo._ID
Some of Android's internals assume the existence of a column whose name
is the value of DBInfo._ID. So in an effort to make the database Android
compliant, I am adding a column with this name. This field is also assumed
to be the table's primary key which is why I am renaming the key column as
opposed to simply adding another column.